ignore duplicate inet6 interfaces
Currently, the debian provider fails with
a malformed error when there are duplciate
interfaces.

This is generally useful, except for the case
where interface definitions exist for ipv4 and
ipv6 interfaces.

In this case, this patch will simply ignore
ipv6 interfaces. While this will probably not
work in the case where a user wants to create
both ipv4/6 interfaces. it will prevent things
from failinng in the limited use case where the
use only cares about ipv4 and happens to have
ipv6 interfaces defined.